Audio Classification project
This is a little audio classification project in which I tried to predict the artist of the sound track.

I have used tiny chunks of audio files from different artists and have tried to predict who is the artist of the particular track and managed to achieve 100% accuracy during model training.